requirements.txt
# Runtime dependencies
fastapi==0.123.10
uvicorn==0.38.0
python-dotenv==1.2.1
pydantic==2.12.5
pydantic_core==2.41.5
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
requests==2.32.5
binance-common==3.2.0
binance-sdk-derivatives-trading-usds-futures==4.0.0
binance-sdk-spot
pandas==2.3.3
pandas-ta==0.4.71b0
numpy==2.2.6
numba==0.61.2
websockets==15.0.1
websocket-client==1.9.0

# Supporting dependencies
aiohttp==3.13.2
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
anyio==4.12.0
starlette==0.50.0
certifi==2025.11.12
charset-normalizer==3.4.4
cryptography==44.0.0
ecdsa==0.19.0
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
tzlocal==5.3.1
dateparser==1.2.2
urllib3==2.6.0
yarl==1.22.0
attrs==25.4.0
frozenlist==1.8.0
multidict==6.7.0
h11==0.16.0
idna==3.11
llvmlite==0.44.0
packaging==25.0
six==1.17.0
tqdm==4.67.1
typing_extensions==4.15.0
regex==2025.11.3
annotated-types==0.7.0
annotated-doc==0.0.4
propcache==0.4.1
pytokens==0.3.0
typing-inspection==0.4.2

# Test dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
